requirements.txt
flask
python-dotenv
line-bot-sdk==1.20.0
gunicorn
